requirements.txt
akshare
pytest
pytest-mock
python-dotenv
openai